requirements.txt
langchain
langgraph
langchain-core
langchain-text-splitters

openai
tiktoken
python-dotenv

chromadb
numpy

pinecone

streamlit
pypdf
python-docx

docling
